Clip to clip-path: circle() and ellipse()
Same root cause as the polygon() fix that precedes it: the paint walker matched
only the shapes it knew and silently dropped the rest, so a circle() or
ellipse() clip did nothing. That is the whole WPT clip-path-circle-* /
clip-path-ellipse-* family — 19 tests in the last run's results — on top of
issue #1544's own clip-path entries. Fixed in Broiler.HTML; this bumps the
pointer and carries the tests.

Both shapes resolve to an ellipse, which the display list can already express:
a rounded rectangle whose box is the ellipse's bounding box and whose corner
radii are the ellipse's radii is that ellipse. So nothing new was needed in
ClipItem or the graphics backends for this one.

ClipPathPolygonTests becomes ClipPathShapeTests now that it covers the shape
grammar rather than one function, and gains the cases the WPT family pins:
circle() in its six spellings, ellipse() in three, the border box as reference
box, a circle percentage resolving against the diagonal rather than a side, a
zero radius clipping everything away, and the unmodelled position form leaving
the element unclipped. The ellipse assertions check every pixel against the
shape equation, skipping only the boundary band, so a bounding-box
approximation could not pass them.
